+++ geodiff/CMakeLists.txt
@@ -142,8 +142,8 @@ SET(geodiff_src
   src/changesetreader.h
   src/changesetutils.cpp
   src/changesetutils.h
-  src/changesetvarint.cpp
-  src/changesetvarint.h
+  src/changesetgetvarint.h
+  src/changesetputvarint.h
   src/changesetwriter.cpp
   src/changesetwriter.h
   src/driver.cpp
